feat: Add the template library for creating a custom message protocol, thread safe queues and an abstrac base class for the client implementation
Add the first rough implementation of the custom networking architecture
using the architecture by javidx9 (youtube), without the complete implementation.
Also the README.md is updated in order to make the structure of the code a little
better understandable. A unit test for the custom message header is written in the
/test folder which uses gtest.

No issue addressed