feat: Add the architecture of the abstract server base class
Add the implementation of the abstract base class of the server implementation and document it within
the README.md. About the first half of the server implementation is done by now. Currently I am working
on the neworking implementation for accepting and denying connections. The abstract base class is
implemented as a template class where the template parameter describes the message type. It is
part of the header based networking library that I write by incooperating the videos from that playlist:
https://www.youtube.com/watch?v=UbjxGvrDrbw

No issue addressed