feat: Create a test_lib executable in order to check if the server templateimplementation works
Create a test_lib executable to check if the server implementation works. Therefor, the connection
handling was outcommented. There were a cycle dependency in the code by creating the OwnedMessage
with a dependency to ConnectionInterface. This is now solved by forward declare the ConnectionInterface
in the OwnedMessage struct implementation and by including the net_message.hpp header in the
connection_net_interface.hpp (instead of including the headers in each other which results in the
described cycle dependency)

No issue addressed