feat: Rearanged the custom client and server example code and finish the implementation of the custom client
Rearanged the custom client and server code to make the project a little more organized and finish the implementation
of the client base class by defining a resolver for the connection to a given url where the server should be.
Right now, we can only use our custom client to talk to the server, since we use a custom (non RFC standazized)
communication protocol that is defined within the net_message.hpp.

No issue addressed