feat: Add a handshake between the client and the server and make the server code more CPU efficient
Add the handshake with a simple encryption function and implement a handshake procedure before
communication can take place. Also the server code is now blocked until an asynchrounous event happens
that pushes a new incoming message to the front or the back of the message queue. By doing this,
the CPU is not always at 100 procent usage since it is listening for new connections. Therfore, a
std::condition_variable is incooperated.

No issue addressed