feat: Add the simple terminal server
Add the final project of the async I/O tutorial from "dens website" to the corresponding
folder of the repo. The code shows how to abstract the I/O parts of the code away
from the server logic and is a good overview of central boost functionallities
for server development. Especially the implementation of services that the server should deliver via
a defined interface of the .attach() method of the server is a good showcase how you can
implement server logic.

No issue addressed