feat: Add the boost::beast tutorials for an asynchronous HTTP GET client.
Add the boost::beast tutorials for an asynchronous HTTP GET client as well as
a synchronous WebSocket client. Also the README.md of the boost::beast introduction
is updated with some theory how the boost::beast library works as well as what
is the difference between a HTTP connection and a WebSocket connection.

No issue addressed